Preserve clipboard state thoughtout the app
