+++ projects/mercury/src/App.js
@@ -17,6 +17,7 @@ import Routes from "./routes/Routes";
 import WorkspaceTopBar from "./common/components/WorkspaceTopBar";
 import {UploadsProvider} from "./common/contexts/UploadsContext";
 import {CollectionsProvider} from "./common/contexts/CollectionsContext";
+import {ClipboardProvider} from './common/contexts/ClipboardContext';
 
 const App = () => {
     const isMounted = useIsMounted();
@@ -47,25 +48,27 @@ const App = () => {
                     <MuiPickersUtilsProvider utils={DateFnsUtils}>
                         <MuiThemeProvider theme={theme}>
                             <UploadsProvider>
-                                <CollectionsProvider>
-                                    <Provider store={store}>
-                                        <ErrorDialog>
-                                            <Router>
-                                                <Layout
-                                                    requiredAuthorization={requiredRole}
-                                                    renderMenu={() => <Menu />}
-                                                    renderMain={() => (
-                                                        <UsersProvider url={users}>
-                                                            <Routes />
-                                                        </UsersProvider>
-                                                    )}
-                                                    renderTopbar={({name}) => <WorkspaceTopBar name={name} />}
-                                                    renderFooter={({id, version}) => <Footer name={id} version={version} />}
-                                                />
-                                            </Router>
-                                        </ErrorDialog>
-                                    </Provider>
-                                </CollectionsProvider>
+                                <ClipboardProvider>
+                                    <CollectionsProvider>
+                                        <Provider store={store}>
+                                            <ErrorDialog>
+                                                <Router>
+                                                    <Layout
+                                                        requiredAuthorization={requiredRole}
+                                                        renderMenu={() => <Menu />}
+                                                        renderMain={() => (
+                                                            <UsersProvider url={users}>
+                                                                <Routes />
+                                                            </UsersProvider>
+                                                        )}
+                                                        renderTopbar={({name}) => <WorkspaceTopBar name={name} />}
+                                                        renderFooter={({id, version}) => <Footer name={id} version={version} />}
+                                                    />
+                                                </Router>
+                                            </ErrorDialog>
+                                        </Provider>
+                                    </CollectionsProvider>
+                                </ClipboardProvider>
                             </UploadsProvider>
                         </MuiThemeProvider>
                     </MuiPickersUtilsProvider>
