+++ projects/mercury/src/common/redux/reducers/cache/jsonLdBySubjectReducers.js
@@ -18,15 +18,6 @@ const updateMetadataReducer = (state = defaultState, action) => {
                     invalidated: true
                 }
             };
-        case actionTypes.UPDATE_COLLECTION_FULFILLED:
-            return {
-                ...state,
-                [action.meta.id]: {
-                    ...state[action.meta.id],
-                    invalidated: true
-                }
-            };
-
         default:
             return state;
     }
